+++ transaction/views.py
@@ -1,19 +1,25 @@
 #encoding: utf-8
 from django.shortcuts import render, redirect
 from cat.models import Cat, BankAccount
-from transaction.models import Transaction
+from transaction.models import Transaction, AutoTransaction
 from datetime import date
+from django.http import HttpResponseRedirect, HttpResponse
 
 
 def addTransaction(request):
+    if not request.user.is_authenticated():
+        return HttpResponseRedirect("/home/")
     t = Transaction()
     t.description = request.POST["description"]
     t.isIncome = request.POST["isIncome"] == '1' 
-    if  not "catId" in request.POST or not "BAId" in request.POST:
-        return redirect('/transaction/')
-    cid = request.POST["catId"]
+    
+    if request.POST["subcatId"]!="":
+        sbcid= request.POST["subcatId"]
+        c = Cat.objects.get(id=sbcid)
+    else:
+        cid = request.POST["catId"]
+        c = Cat.objects.get(id=cid)
     bid = request.POST["BAId"]
-    c = Cat.objects.get(id=cid)
     ba = BankAccount.objects.get(id=bid)
     t.Category = c
     t.bankAccount = ba
@@ -24,7 +30,7 @@ def addTransaction(request):
         t.cost = request.POST["cost"]
         t.save()
     except :
-        return transaction(request, {'error':'مبلغ را به عدد وارد کنید.', 'error-cost':'1'})
+        return transaction(request, {'error':'مبلغ را به عدد وارد کنید.', 'error_cost':'1'})
     
     try:
         t.date = request.POST["date"]
@@ -32,7 +38,7 @@ def addTransaction(request):
         t.save()
     except:
         t.delete()
-        return transaction(request, {'error':'فرمت تاریخ نادرست است. YYYY-MM-DD', 'error-date':'1'})
+        return transaction(request, {'error':'فرمت تاریخ نادرست است. YYYY-MM-DD', 'error_date':'1'})
     
     
     
@@ -40,6 +46,8 @@ def addTransaction(request):
     return redirect('/transaction/')
 
 def transaction(req, extra={}):
+    if not req.user.is_authenticated():
+        return HttpResponseRedirect("/home/")
     if req.method == "POST" and 'formID' in req.POST:
         if req.POST['formID'] == "1":
             new = Cat(name=req.POST['name'], isSub=False, parentCat=None, user=req.user)
@@ -54,12 +62,15 @@ def transaction(req, extra={}):
             p.name = req.POST['new']
             p.save()
     T = Transaction.objects.all().filter(user=req.user).order_by('-date')
-    a = {"Tran":T, 'cats': Cat.objects.filter(isSub=False , user=req.user), 'bankAccounts': BankAccount.objects.filter(user=req.user)}
+    TT = AutoTransaction.objects.all().filter(user=req.user).order_by('-date')
+    a = {"Tran":T, "AutoTran": TT, 'cats': Cat.objects.filter(isSub=False , user=req.user), 'bankAccounts': BankAccount.objects.filter(user=req.user)}
     a.update(extra)
     return  render(req, 'transaction.html', a)
 
 
 def deleteTransaction(request):
+    if not request.user.is_authenticated():
+        return HttpResponseRedirect("/home/")
     id = request.POST["id"]
     T = Transaction.objects.get(id=id)
     T.delete()
@@ -68,6 +79,8 @@ def deleteTransaction(request):
 
 
 def editTransaction(request):
+    if not request.user.is_authenticated():
+        return HttpResponseRedirect("/home/")
     id = request.POST["id"]
     t = Transaction.objects.get(id=id)
     t.date = request.POST["date"]
@@ -84,16 +97,125 @@ def editTransaction(request):
     t.save()
     return redirect('/transaction/')
 def bankAccountAdd (req):
+    if not req.user.is_authenticated():
+        return HttpResponseRedirect("/home/")
     new = BankAccount(name=req.POST['name'], user=req.user)
     new.save()
     return redirect('/transaction/') 
 def bankAccountDel (req):
+    if not req.user.is_authenticated():
+        return HttpResponseRedirect("/home/")
     BankAccount.objects.filter(id=req.POST['id']).delete()
     return redirect('/transaction/')
 def bankAccountEdit (req):
+    if not req.user.is_authenticated():
+        return HttpResponseRedirect("/home/")
     p = BankAccount.objects.get(id=req.POST['id'])
     p.name = req.POST['new']
     p.save()
     return redirect('/transaction/')
 
 
+
+
+def addAutoTransaction(request):
+    if not request.user.is_authenticated():
+        return HttpResponseRedirect("/home/")
+    print 1
+    t = AutoTransaction()
+    print 2
+    t.description = request.POST["description"]
+    print 3
+    t.isIncome = request.POST["isIncome"] == '1'
+    print 4
+    if request.POST["subcatId"]!="":
+        sbcid= request.POST["subcatId"]
+        c = Cat.objects.get(id=sbcid)
+    else:
+        cid = request.POST["catId"]
+        c = Cat.objects.get(id=cid)
+    print 5
+    bid = request.POST["BAId"]
+    ba = BankAccount.objects.get(id=bid)
+    t.Category = c
+    t.bankAccount = ba
+    t.user = request.user
+    t.date=date.today()
+    t.lastModified = date.today()
+
+    t.interval = request.POST["interval"]
+    print 6
+
+    try:
+        t.cost = request.POST["cost"]
+        t.save()
+    except :
+        return transaction(request, {'auto_error':'مبلغ را به عدد وارد کنید.', 'auto_error_cost':'1'})
+
+    print 7
+    try:
+        t.date = request.POST["date"]
+        t.lastModified = None
+        print t.date
+        t.save()
+    except:
+        t.delete()
+        return transaction(request, {'auto_error':'فرمت تاریخ نادرست است. YYYY-MM-DD', 'auto_error_date':'1'})
+
+    t.save()
+    return redirect('/transaction/')
+
+def editAutoTransaction(request):
+    if not request.user.is_authenticated():
+        return HttpResponseRedirect("/home/")
+    id = request.POST["id"]
+    t = AutoTransaction.objects.get(id=id)
+    t.interval = request.POST["interval"]
+    t.date = request.POST["date"]
+    t.description = request.POST["description"]
+    t.cost = request.POST["cost"]
+    if request.POST["isIncome"] == '1':
+      t.isIncome = True
+    else:
+      t.isIncome = False
+    c = Cat.objects.get(id=request.POST["catId"])
+    ba = BankAccount.objects.get(id=request.POST['BAId'])
+    t.Category = c
+    t.bankAccount = ba
+    t.save()
+    return redirect('/transaction/')
+
+def deleteAutoTransaction(request):
+    if not request.user.is_authenticated():
+        return HttpResponseRedirect("/home/")
+    id = request.POST["id"]
+    T = AutoTransaction.objects.get(id=id)
+    T.delete()
+    print "@"*20
+    return redirect('/transaction/')
+
+
+
+def autoTransCheck(request):
+    const = dict()
+    const[1] = 1 ; const[2] = 7 ; const[3] = 30 ;
+
+    for tran in AutoTransaction.objects.all():
+        if date.today() >= tran.date:
+            if tran.lastModified == None or (date.today()-tran.lastModified).days >= const[tran.interval]:
+                print '------------------'
+                tran.lastModified = date.today()
+
+                t = Transaction()
+                t.description = u'خودکار - '+ tran.description
+                t.isIncome = tran.isIncome
+                t.date = date.today()
+                t.Category = tran.Category
+                t.bankAccount = tran.bankAccount
+                t.user = tran.user
+                t.cost = tran.cost
+                t.save()
+
+                tran.save()
+
+    return redirect('/transaction/')
